requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pydantic
google-genai
python-dotenv
dnspython